Restructure tests, import optimize & reformat
