--- tests/BAC0/read-point-bacpypes.py
@@ -1,3 +1,0 @@
-from device_test_settings import  test_device_url, net_url, test_device
-
-
