+++ tests/bacnet/bac0/bac0-logging.py
@@ -1,6 +1,7 @@
 import BAC0
 import logging
 import sys
+
 # turn off BAC0 logging
 BAC0.log_level('silence')
