+++ tests/bacnet/bac0/bac0_points.py
@@ -1,13 +1,6 @@
-
-
-
-
-
 import BAC0
 
-from device_test_settings import  test_device_url, net_url, test_device
-
-
+from tests.bacnet.bac0.device_test_settings import test_device_url, net_url, test_device
 
 bacnet = BAC0.lite(ip=net_url)
 print(test_device_url)
@@ -19,16 +12,13 @@ obj = "analogInput"
 obj_instance = "1"
 prop = "presentValue"
 
-
-device = BAC0.device(test_device_url, test_device,bacnet,poll=0, history_size=1)
+device = BAC0.device(test_device_url, test_device, bacnet, poll=0, history_size=1)
 print(device.points)
-
-
 
 find_overrides = device.find_overrides()
 print(find_overrides)
 # https://bac0.readthedocs.io/en/latest/_modules/BAC0/core/devices/Points.html?highlight=Is_overridden
-is_overridden= device['AO 2'].__getitem__("presentValue")
+is_overridden = device['AO 2'].__getitem__("presentValue")
 print(is_overridden)
 # device.release_all_overrides()
 # device['point'].is_overridden
