+++ tests/bacnet/bac0/bac0_read.py
@@ -1,8 +1,6 @@
 import BAC0
 
-from device_test_settings import  test_device_url, net_url, test_device
-
-
+from tests.bacnet.bac0.device_test_settings import test_device_url, net_url, test_device
 
 bacnet = BAC0.lite(ip=net_url)
 print(test_device_url)
@@ -14,8 +12,7 @@ obj = "analogInput"
 obj_instance = "1"
 prop = "presentValue"
 
-
 # device = BAC0.device(test_device_url, test_device,bacnet,poll=0, history_size=1)
-readObj  = f"{test_device_url} {obj} {obj_instance}  {prop}"
+readObj = f"{test_device_url} {obj} {obj_instance}  {prop}"
 read = bacnet.read(readObj)
 print(read)
