+++ tests/bacnet/bac0/device_test_settings.py
@@ -1,8 +1,5 @@
-
-
-
 network = {
-    'network_ip':'192.168.0.100',
+    'network_ip': '192.168.0.100',
     'network_mask': 24,
     'network_port': 47808,
     'network_device_id': 22,
@@ -12,11 +9,8 @@ network = {
 
 net_url = f"{network['network_ip']}/{network['network_mask']}:{network['network_port']}"
 
-
-
-
 test_device = {
-    'network_ip':'192.168.0.202',
+    'network_ip': '192.168.0.202',
     'network_mask': 24,
     'network_port': 47808,
     'network_device_id': 202,
@@ -25,10 +19,4 @@ test_device = {
 
 test_device_url = f"{test_device['network_ip']}/{test_device['network_mask']}:{test_device['network_port']}"
 
-
-
 test_device_id = test_device['network_device_id']
-
-
-
-
