+++ tests/bacnet/bac0/test-bacnet-server.py
@@ -1,17 +1,7 @@
 ##The following code connects to a DXR and then command damper position open adn readys flow back
 # The following modules need to be enabled via [pip python "module"]
-from collections import namedtuple
 
-import BAC0
-from BAC0.core.devices.create_objects import create_AV, create_AI, _make_mutable
-from bacpypes.basetypes import EngineeringUnits
-
-# hvac = BAC0.lite(port=47808)
-# temp = create_AV(oid=1, name="temp", pv=12.8, pv_writable=True)
-# temp.units = EngineeringUnits.enumerations["degreesCelsius"]
-# temp.description = "AVG Temp"
-# hvac.this_application.add_object(temp)
-
+import time
 
 # hvac = BAC0.lite(port=47801)
 # temp = create_AV(oid=1, name="temp", pv=12.8, pv_writable=True)
@@ -33,14 +23,16 @@ from bacpypes.basetypes import EngineeringUnits
 # params.test_device.disconnect()
 # params.bacnet.disconnect()
 import BAC0
-import random
-import time
-from BAC0.core.devices.create_objects import create_AI
-from collections import namedtuple
+from BAC0.core.devices.create_objects import create_AV
 from BAC0.tasks.RecurringTask import RecurringTask
 from bacpypes.primitivedata import Real
 
-from BAC0.core.devices.create_objects import create_AI
+
+# hvac = BAC0.lite(port=47808)
+# temp = create_AV(oid=1, name="temp", pv=12.8, pv_writable=True)
+# temp.units = EngineeringUnits.enumerations["degreesCelsius"]
+# temp.description = "AVG Temp"
+# hvac.this_application.add_object(temp)
 
 
 def changeValueOfMyAI():
@@ -49,9 +41,10 @@ def changeValueOfMyAI():
     value = value + 1 if value < 100 else 0
     obj.presentValue = Real(value)
 
-ip='192.168.0.100/24:47808'
-deviceId =100
-localObjName= "BAC0_Fireplace"
+
+ip = '192.168.0.100/24:47808'
+deviceId = 100
+localObjName = "BAC0_Fireplace"
 
 # ip = None,
 # port = None,
@@ -61,7 +54,7 @@ localObjName= "BAC0_Fireplace"
 # ping = True,
 
 
-bacnet = BAC0.lite(ip=ip, deviceId=deviceId, localObjName= localObjName)
+bacnet = BAC0.lite(ip=ip, deviceId=deviceId, localObjName=localObjName)
 # bacnet = BAC0.lite()
 # bacnet = BAC0.lite(ip='192.168.0.100/24:47808')
 bacnet.this_application.add_object(create_AV(oid=0, name='MyAI', pv=Real(0), pv_writable=True))
