+++ tests/bacnet/misc/BACpypes.ini
@@ -1,0 +1,7 @@
+[BACpypes]
+objectName : Betelgeuse
+address : 192.168.0.100/24
+objectIdentifier : 599
+maxApduLengthAccepted : 1024
+segmentationSupported : segmentedBoth
+vendorIdentifier : 15
