--- tests/misc examples/BACpypes.ini
@@ -1,7 +1,0 @@
-[BACpypes]
-objectName: Betelgeuse
-address: 192.168.0.100/24
-objectIdentifier: 599
-maxApduLengthAccepted: 1024
-segmentationSupported: segmentedBoth
-vendorIdentifier: 15
